requirements.txt
django 
djangorestframework
psutil
pyinstaller